Module Ark_bls12_377_g1

Bindings to the g1 modulue in the ark-bls12-377 rust package. This is used * as a reference implemenetation to test for correctness. * * Coordinates here are affine coordinates in the regular space (specifically, * not montgomery space). The underlying rust implementation uses montgomery * space, this is hidden by this module altogether. *

type affine

val sexp_of_affine : affine -> Sexplib0.Sexp.t

val equal_affine : affine -> affine -> bool

val create : x:Z.t -> y:Z.t -> infinity:bool -> affine

Constructs an arbitrary value in the finite field. Note that an arbitrarily * constructed point may not lie on the curve. *

val add : affine -> affine -> affine

Performs a point-addition between two points. This operation is * commutative.

val neg : affine -> affine

If non infinity, then return (x, -y). Otherwise, returns infinity.

val mul : affine -> by:int -> affine

Given a point p, computes (p * by) * * Raises an exception if by is negative.

val is_on_curve : affine -> bool

Returns true if the point is on the eliptic bls12-377 curve. Note that for * arbitrarily constructed points via create, this might not be the case.

val x : affine -> Z.t

Returns the x-coordinates of the affine point.

val y : affine -> Z.t

Returns the y-coordinates of the affine point.

val infinity : affine -> bool

Returns true if the point is in infinity.

Curve Constants

* * The curve is of the form y^2 = x^3 + ax + b, where y and x are in the finite * field modulus. Note that the values are cached to prevent repeated FFI * calls into rust where necessary.

val coeff_a : unit -> Z.t

Returns a

val coeff_b : unit -> Z.t

Returns b

val modulus : unit -> Z.t

Returns the prime modulus.

val subgroup_generator : unit -> affine

Returns the bls12-377 subgroup generator.

val mul_wide : part_width:int -> affine -> by:Hardcaml.Bits.t -> affine

Multiply by an arbitrary width Hardcaml Bits vector.

For efficiency part_width should be as wide as possible which is 61 bits in this implementation.

module For_testing : sig ... end