Module Elliptic_curve_lib.Ec_fpn_mixed_add

Fully unrolled eliptic curve point adding.

Based on the formula here:

https://en.wikibooks.org/wiki/Cryptography/Prime_Curve/Jacobian_Coordinates#Point_Addition_(12M_+_4S)

Currently only mixed Jacobian and Affine point addition is supported.

An optimization requires the previous stage provide this module with data_in1.z^2 as an input.

In order to reduce the amount of DSPs used we arbitrate multiplications and squares in certain stages.

module Config = Ec_fpn_ops_config

val latency : Config.t -> int

Latency of the eliptic curve point adding datapath.

module With_interface (M : sig ... end) : sig ... end