Module Adder_subtractor_pipe.O

Output of the adder/subtractor with carry/borrow bits.

When instantiating a pure adder, the carry will be the carry bits with width = Int.ceil_log2 num_operands. For a subtractor, it will the borrow bits. For a mixed adder/subtractor implementation, you can treat carry @: result as your signed final result.

The width of result matches the input operands, the width of carry is dependent on the exact operands and number of operands.

type 'a t = {

result : 'a;

carry : 'a;

}

val carry : 'a t -> 'a

val result : 'a t -> 'a

module Fields : sig ... end