Module Config.Level

type t = {

radix : Radix.t;	(*
Specified the number of splitting to be done on terms when building the multiplier.
*)

pre_adder_stages : int;	(*
The number of pipeline stages of the pre multiplication adder. For a multiplication of 2 N-bit numbers, this is addition of width ceil(N/2)
*)

middle_adder_stages : int;	(*
The number of pipeline stages of the middle adder. This is an adder of width round_up_to_multiple_of_2(N)
*)

post_adder_stages : int;	(*
The number of pipeline stages on the final adder. For a multiplication of 2 N-bit numbers, this will be the number of pipeline stages for the final 2N-bit adder.
*)

}

val sexp_of_t : t -> Sexplib0.Sexp.t