Module Msm_pippenger.Msm_result_to_host

Takes in points from the MSM block and packs them input a 512bit stream where every coordinate and point has been aligned to 64 bits, and can then be sent up to a central controller and back to the host.

Optionally drops or sends the t coordinate.

module Make (Config : Config.S) : sig ... end