Module Msm_pippenger.Scalar_transformation

A pipeline that transforms scalars into the form that is expected on the FPGA. The main transform used currently is converting scalars from the range 1, 2^c-1 into the range 2^c-1, -2^c-1 which allows for mapping negative points into positive point buckets.

module Axi512 = Hardcaml_axi.Axi512

module Make (Config : Config.S) : sig ... end