Parameter Make.1-M

val centre_slr : int option

val partitions : Partition.t list

val data_bits : int

val ram_read_latency : int

The underlying XPM's read latency

val ram_lookup_latency : int

The latency on the inputs to the ram (read enables, read_window, write * enable and address

val ram_output_latency : int

Additional pipelining on the read data for it to cross SLR gracefully.