Module Twisted_edwards_lib

module Adder_intf : sig ... end

module Config : sig ... end

module Mixed_add : sig ... end

Fully pipelined/half-pipelined mixed addition for twisted edwards curve with a = -1

module Mixed_add_precompute : sig ... end

Fully/Half pipelined mixed addition for a scaled twisted edwards curve curve (ie: a = -1), optimized with precomputation.

module Num_bits : sig ... end

module Xyt : sig ... end

module Xyzt : sig ... end