Parameter Make.1-Config

include Hardcaml_ntt.Core_config.S

val logn : Base.int

Log (base 2) of core transform size.

For the 4 step algorithm, this is half the total transform size.

val support_4step_twiddle : Base.bool

Enable support for the twiddle pass used in the 4 step algorithm.

Configuration of multiple core instantiations

val logcores : Base.int

Log (base 2) number of cores to instantiate per block. Each block has a single controller component.

val logblocks : Base.int

Log (base 2) of number of blocks of cores to instantiate.

val memory_layout : Memory_layout.t